refactor: migrate to feature-based directory structure (Phase 2)
Phase 2 complete: ディレクトリ構造変更 (Feature-based migration)

Changes:
- Reorganized codebase from type-based to feature-based structure
- Created features/ directory for domain-specific code:
  - features/auth/: Authentication components, contexts, and pages
  - features/search/: Search-related components, hooks, and types
  - features/history/: History page (will house history feature in Phase 3)
- Created shared/ directory for reusable code:
  - shared/components/layouts/: AuthLayout and GuestLayout
  - shared/hooks/: useStorageCache
  - shared/utils/: storage and fileValidation utilities
- Updated all import paths to reflect new structure
- Fixed test mocks to use correct relative paths
- Removed duplicate AuthProvider wrap in main.tsx

Structural integrity:
- All tests passing (91/91) - behavior unchanged
- Used git mv to preserve file history
- No functional changes - pure refactoring

This follows Tidy First principles: structural changes separated from
behavioral changes, validated by test suite.
