+++ frontend/src/features/auth/components/LoginForm/LoginForm.tsx
@@ -1,5 +1,5 @@
 import { useCallback, useId, useState } from "react";
-import { useAuth } from "../contexts/AuthContext";
+import { useAuth } from "../../contexts/AuthContext";
 import "./LoginForm.css";
 
 interface LoginFormProps {
