+++ frontend/src/features/auth/pages/LoginPage.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "../components/LoginForm";
+import { LoginForm } from "../components/LoginForm/LoginForm";
 import "./LoginPage.css";
 
 export default function LoginPage() {
