+++ frontend/src/features/search/components/FileUploadArea.tsx
