+++ frontend/src/features/search/hooks/useFileUpload.ts
@@ -1,6 +1,6 @@
 import { useCallback, useState } from "react";
 import type { UploadedFile } from "../types/file";
-import { validateFiles } from "../utils/fileValidation";
+import { validateFiles } from "../../../shared/utils/fileValidation";
 
 export interface UseFileUploadReturn {
   files: UploadedFile[];
