+++ frontend/src/features/search/types/searchResult.ts
