+++ frontend/src/shared/components/layouts/AuthLayout.test.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import GuestLayout from "./GuestLayout";
+import AuthLayout from "./AuthLayout";
 
 // Mock react-router-dom components
 vi.mock("react-router-dom", () => ({
@@ -11,9 +11,9 @@ vi.mock("react-router-dom", () => ({
 }));
 
 /**
- * Mock Supabase client to isolate GuestLayout from external dependencies.
+ * Mock Supabase client to isolate AuthLayout from external dependencies.
  */
-vi.mock("../lib/supabaseClient", () => ({
+vi.mock("../../../lib/supabaseClient", () => ({
   supabase: {
     auth: {
       getSession: vi.fn(),
@@ -30,40 +30,52 @@ const mockUseAuth = vi.fn();
 /**
  * Mock AuthContext with configurable useAuth return value.
  */
-vi.mock("../contexts/AuthContext", () => ({
+vi.mock("../../../features/auth/contexts/AuthContext", () => ({
   useAuth: () => mockUseAuth(),
 }));
 
-describe("GuestLayout", () => {
+describe("AuthLayout", () => {
   beforeEach(() => {
     mockUseAuth.mockClear();
   });
 
-  describe("when user is not authenticated", () => {
-    it("should render children via Outlet", () => {
-      mockUseAuth.mockReturnValue({
-        user: null,
-        loading: false,
-      });
-
-      const { getByTestId } = render(<GuestLayout />);
-
-      expect(getByTestId("outlet")).toBeInTheDocument();
-    });
-  });
-
   describe("when user is authenticated", () => {
-    it("should redirect to /", () => {
+    it("should render children via Outlet", () => {
       mockUseAuth.mockReturnValue({
         user: { id: "test-user", email: "test@example.com" },
         loading: false,
       });
 
-      const { getByTestId } = render(<GuestLayout />);
+      const { getByTestId } = render(<AuthLayout />);
+
+      expect(getByTestId("outlet")).toBeInTheDocument();
+    });
+
+    it("should apply auth-layout class", () => {
+      mockUseAuth.mockReturnValue({
+        user: { id: "test-user", email: "test@example.com" },
+        loading: false,
+      });
+
+      const { container } = render(<AuthLayout />);
+
+      const authLayoutDiv = container.querySelector(".auth-layout");
+      expect(authLayoutDiv).toBeInTheDocument();
+    });
+  });
+
+  describe("when user is not authenticated", () => {
+    it("should redirect to /login", () => {
+      mockUseAuth.mockReturnValue({
+        user: null,
+        loading: false,
+      });
+
+      const { getByTestId } = render(<AuthLayout />);
 
       const navigate = getByTestId("navigate");
       expect(navigate).toBeInTheDocument();
-      expect(navigate.textContent).toContain("Redirect to /");
+      expect(navigate.textContent).toContain("/login");
     });
   });
 
@@ -74,7 +86,7 @@ describe("GuestLayout", () => {
         loading: true,
       });
 
-      const { getByText } = render(<GuestLayout />);
+      const { getByText } = render(<AuthLayout />);
 
       expect(getByText("認証確認中...")).toBeInTheDocument();
     });
@@ -85,7 +97,7 @@ describe("GuestLayout", () => {
         loading: true,
       });
 
-      const { container } = render(<GuestLayout />);
+      const { container } = render(<AuthLayout />);
 
       const loadingDiv = container.querySelector(".loading");
       expect(loadingDiv).toBeInTheDocument();
