+++ frontend/src/shared/components/layouts/AuthLayout.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from "react-router-dom";
-import { useAuth } from "../contexts/AuthContext";
+import { useAuth } from "../../../features/auth/contexts/AuthContext";
 import "./AuthLayout.css";
 
 /**
