+++ frontend/src/shared/hooks/useStorageCache.ts
