+++ frontend/src/shared/utils/storage.ts
